adapt to htslib 1.10 (#103)
* adapt to htslib 1.10

* fixed record creating in test case

* testcase for issue #101

* always report positive svlen in order to work around issue https://github.com/samtools/bcftools/issues/874

* use proper rust-htslib dependency

* updated dependencies

+++ src/calling/cnvs.rs
@@ -303,7 +303,7 @@ impl<'a> CNVCall<'a> {
         contig_len: u32,
     ) -> Result<()> {
         record.set_rid(Some(rid));
-        record.set_pos(self.pos as i32);
+        record.set_pos(self.pos as i64);
         record.set_alleles(&[b"N", b"<CNV>"])?;
         record.push_info_integer(b"END", &[self.end as i32])?;
         record.push_info_integer(b"SVLEN", &[self.len() as i32])?;
